backend/requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.1.1
firebase-admin==6.5.0
dotenv==0.9.9